lift fov functions into namespace fov
