+++ src/map.hpp
@@ -37,27 +37,6 @@ struct Tile {
 	}
 };
 
-enum class FovType { SQUARE, CIRCLE };
-
-class Shadow {
-public:
-	Shadow(float start, float end) : start(start), end(end) {;}
-	float start;
-	float end;
-
-	static Shadow projectTile(int row, int column);
-	bool contains(Shadow other);
-
-};
-
-class ShadowLine {
-public:
-	std::vector<Shadow> shadows;
-
-	bool isInShadow(Shadow projection);
-	bool isFullShadow() { return shadows.size() == 1 && shadows[0].start == 0 && shadows[0].end == 1; }
-	void addShadow(Shadow shadow);
-};
 
 enum class BreakDirection { HORIZONTAL, VERTICAL };
 enum class MapType { BUILDING, WATER, PILLARS };
@@ -81,13 +60,10 @@ public:
 	bool canWalk(int x, int y) const;
 	bool isInFov(int x, int y) const;
 	bool isExplored(int x, int y) const;
-	void computeFov(int x, int y, int radius = constants::DEFAULT_FOV_RADIUS, FovType fovType = FovType::CIRCLE);
-	void computeFovForOctant(int x, int y, int octant, int radius = constants::DEFAULT_FOV_RADIUS, FovType fovType = FovType::CIRCLE);
 	void setWorld(World* w) { world = w; }
 
 private:
 	World* world;
-	Vec<int> transformOctant(int row, int col, int octant);
 	friend class boost::serialization::access;
 	template<class Archive>
 	void serialize(Archive & ar, const unsigned int version) {
