+++ src/world.hpp
@@ -3,6 +3,7 @@
 
 #include <vector>
 #include "map.hpp"
+#include "fov.hpp" //FIXME move implementation to cpp
 
 class Actor;
 struct Event;
@@ -26,7 +27,7 @@ public:
 	std::vector<Actor*> getActorsAt(int x, int y);
 
 
-	void computeFov(int x, int y, float r) { map.computeFov(x, y, r); }
+	void computeFov(int x, int y, float r) { fov::computeFov(&map, x, y, r); }
 	bool isWall(int x, int y) { return map.isWall(x, y); }
 	bool canWalk(int x, int y);
 	bool isInFov(int x, int y) { return map.isInFov(x, y); }
